// src/utils/types/auth.types.ts
export const enum AccessScopes {
  SIGN_UP = "SIGN_UP",
  LOGGED_IN = "LOGGED_IN",
  ARTIST_CREATE = "ARTIST_CREATE",
  ARTIST_TEAM_ADD_MEMBERS = "ARTIST_TEAM_ADD_MEMBERS",
  ARTIST_TEAM_INVITATION_CONFIRM = "ARTIST_TEAM_INVITATION_CONFIRM",
  PROFILE_PAGE_CREATE = "PROFILE_PAGE_CREATE",
  PROFILE_EDIT = "PROFILE_EDIT",
  PROFILE_EDIT_ADVANCED = "PROFILE_EDIT_ADVANCED",
  MODERATOR_ROLE_TRANSFER = "MODERATOR_ROLE_TRANSFER",
  FORUM_POST_CREATE = "FORUM_POST_CREATE",
  FORUM_POST_DELETE = "FORUM_POST_DELETE",
  FORUM_WRITEING = "FORUM_WRITEING",
  FORUMS_READING = "FORUMS_READ",
  FORUM_SEARCHING = "FORUM_SEARCH",
  ARTISTS_LIST_WATCHING = "ARTISTS_LIST_WATCH",
  ARTIST_SUBSCRIBE = "ARTIST_SUBSCRIBE",
  USER_VERIFY = "USER_VERIFY",
  ROLE_CHOSE_CREATIVE_OR_FUN = "ROLE_CHOSE_CREATIVE_OR_FUN",
}